+++ frontend/src/App.vue
@@ -40,19 +40,7 @@ export default {
 	name: 'App',
 	methods: {
 		main: function () {
-			if (this.$router.history.current.path != '/') {
-				this.$router.push('/');
-			}
-		},
-		login: function () {
-			if (this.$router.history.current.path != '/login') {
-				this.$router.push('/login');
-			}
-		},
-		post: function () {
-			if (this.$router.history.current.path != '/') {
-				this.$router.push('/');
-			}
+			this.$router.push('/');
 		},
 		board(category) {
 			this.$router.push(`/board/${category}`);